+++ frontend/src/players_view/PlayersView.css
