+++ frontend/src/table_components/SearchBar.jsx
