+++ frontend/src/table_components/StatsTable.css
