+++ frontend/src/table_components/StatsTable.jsx
